Inline onEvent function in DomainStats
