+++ src/groups/mqb/mqbstat/mqbstat_domainstats.h
@@ -28,6 +28,7 @@
 // exposing methods to initialize the stat contexts.
 
 // MQB
+#include <bmqst_statcontext.h>
 
 // BDE
 #include <bsl_memory.h>
@@ -70,6 +71,16 @@ class DomainStats {
     struct Stat {
         // TYPES
         enum Enum { e_CFG_MSGS, e_CFG_BYTES, e_QUEUE_COUNT };
+    };
+
+    //------------------------
+    // struct DomainStatsIndex
+    //------------------------
+
+    /// Namespace for the constants of stat values that applies to the queues
+    /// from the clients
+    struct DomainStatsIndex {
+        enum Enum { e_STAT_CFG_MSGS, e_STAT_CFG_BYTES, e_STAT_QUEUE_COUNT };
     };
 
   private:
@@ -115,7 +126,8 @@ class DomainStats {
     /// Update statistics for the event of the specified `type` and with the
     /// specified `value` (depending on the `type`, `value` can represent
     /// the number of bytes, a counter, ...
-    void onEvent(EventType::Enum type, bsls::Types::Int64 value);
+    template <EventType::Enum type>
+    void onEvent(bsls::Types::Int64 value);
 
     /// Return a pointer to the statcontext.
     bmqst::StatContext* statContext();
@@ -148,6 +160,30 @@ struct DomainStatsUtil {
 inline bmqst::StatContext* DomainStats::statContext()
 {
     return d_statContext_mp.get();
+}
+
+template <>
+inline void DomainStats::onEvent<DomainStats::EventType::Enum::e_CFG_MSGS>(
+    bsls::Types::Int64 value)
+{
+    BSLS_ASSERT_SAFE(d_statContext_mp && "initialize was not called");
+    d_statContext_mp->setValue(DomainStatsIndex::e_STAT_CFG_MSGS, value);
+}
+
+template <>
+inline void DomainStats::onEvent<DomainStats::EventType::Enum::e_CFG_BYTES>(
+    bsls::Types::Int64 value)
+{
+    BSLS_ASSERT_SAFE(d_statContext_mp && "initialize was not called");
+    d_statContext_mp->setValue(DomainStatsIndex::e_STAT_CFG_BYTES, value);
+}
+
+template <>
+inline void DomainStats::onEvent<DomainStats::EventType::Enum::e_QUEUE_COUNT>(
+    bsls::Types::Int64 value)
+{
+    BSLS_ASSERT_SAFE(d_statContext_mp && "initialize was not called");
+    d_statContext_mp->setValue(DomainStatsIndex::e_STAT_QUEUE_COUNT, value);
 }
 
 }  // close package namespace
